adds more tests for near-me, reorganizes tests
